Added basis support for individual model builder retrieval + cleaned and expanded test environments
