+++ django_elasticsearch_model_binder/__init__.py
@@ -1,6 +1,5 @@
-from django_elasticsearch_model_binder.mixins import (
-    ESModelBinderMixin, ESQuerySetMixin
-)
+from django_elasticsearch_model_binder.mixins import ESQuerySetMixin
+from django_elasticsearch_model_binder.models import ESBoundModel
 from django_elasticsearch_model_binder.utils import ExtraModelFieldBase
 
-__all__ = ['ESModelBinderMixin', 'ESQuerySetMixin', 'ExtraModelFieldBase']
+__all__ = ['ESBoundModel', 'ESQuerySetMixin', 'ExtraModelFieldBase']
