+++ tests/test_app/models.py
@@ -1,12 +1,12 @@
 from django.db import models
 
-from django_elasticsearch_model_binder import ESModelBinderMixin
+from django_elasticsearch_model_binder import ESBoundModel
 
 from tests.test_app.managers import ESEnabledQuerySet
 from tests.test_app.utils import UniqueIdentiferField
 
 
-class User(ESModelBinderMixin, models.Model):
+class User(ESBoundModel):
     email = models.EmailField(max_length=254)
 
     es_cached_extra_fields = (UniqueIdentiferField,)
@@ -14,7 +14,7 @@ class User(ESModelBinderMixin, models.Model):
     objects = ESEnabledQuerySet.as_manager()
 
 
-class Author(ESModelBinderMixin, models.Model):
+class Author(ESBoundModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     publishing_name = models.CharField(max_length=25, blank=True, null=True)
     age = models.IntegerField()
