+++ tox.ini
@@ -1,10 +1,11 @@
 [tox]
 envlist =
-    {py35,py34}-django-{111}-{es6}
+    {py38,py37,py36,py35,py34}-django-{30,22,111}-{es6}
 
 [testenv]
 setenv =
     PYTHONPATH = {toxinidir}
+    DJANGO_SETTINGS_MODULE = tests.test_app.settings
 commands = pytest
 deps =
     django
@@ -14,6 +15,5 @@ deps =
     typing
 
 [pytest]
-DJANGO_SETTINGS_MODULE = settings
 python_files = tests.py test_*.py *_tests.py
 django_find_project = false
